intermediate renaming (case issue)
