+++ packages/extras/src/lib/index.ts
@@ -22,8 +22,8 @@ export { default as Float } from './components/Float/Float.svelte'
 export { default as GLTF } from './components/GLTF/GLTF.svelte'
 export { default as Gizmo } from './components/Gizmo/Gizmo.svelte'
 export { default as ContactShadows } from './components/ContactShadows/ContactShadows.svelte'
-export { default as CubeEnvironment } from './components/environment/CubeEnvironment.svelte'
-export { default as Environment } from './components/environment/Environment.svelte'
+export { default as CubeEnvironment } from './components/x/CubeEnvironment.svelte'
+export { default as Environment } from './components/x/Environment.svelte'
 export { default as ImageMaterial } from './components/ImageMaterial/ImageMaterial.svelte'
 export { default as Grid } from './components/Grid/Grid.svelte'
 export { default as RoundedBoxGeometry } from './components/RoundedBoxGeometry/RoundedBoxGeometry.svelte'
